+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.4.0 (2024-04-04)
 
 - adds `onLoadingAnimationCompleted` and `animating` parameter to `CircularWidgetLoading`
+- adds `index` to `DotBuilder`
 
 ## 0.3.2 (2022-06-13)